fix: address PR #2106 CI failures — move Linux-only tests to build-constrained file (Issue #2095)
`TestLoadOrGenerateFallbackMachineID` and `TestPlatformEncryptor_FallbackMachineID`
referenced `loadOrGenerateFallbackMachineID` and `fallbackMachineIDFile` which are
defined only in `crypto_linux.go` (//go:build linux). Placing them in the
cross-platform `crypto_test.go` caused compile failures on macOS and Windows.

Moved both tests into a new `crypto_linux_test.go` file (//go:build linux), removed
the now-redundant runtime `t.Skip` guards, and dropped the unused cross-platform
test-file references to Linux-only symbols.
