+++ pkg/secrets/providers/steward/crypto_test.go
@@ -162,57 +162,3 @@ func TestAesGcmEncryptor_ShortCiphertext(t *testing.T) {
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "ciphertext too short")
 }
-
-// TestPlatformEncryptor_FallbackMachineID verifies that the platform encryptor works
-// when /etc/machine-id is absent or empty (the case in containers without systemd).
-// This test only runs on Linux because the fallback is Linux-specific.
-func TestPlatformEncryptor_FallbackMachineID(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("fallback machine-id is Linux-only")
-	}
-	tmpDir := t.TempDir()
-
-	enc, err := newPlatformEncryptor(tmpDir)
-	require.NoError(t, err, "newPlatformEncryptor must succeed even without /etc/machine-id")
-	require.NotNil(t, enc)
-
-	// Verify round-trip works with the fallback key.
-	plaintext := []byte("test secret for fallback key")
-	ciphertext, err := enc.Encrypt(plaintext)
-	require.NoError(t, err)
-	assert.NotEmpty(t, ciphertext)
-
-	decrypted, err := enc.Decrypt(ciphertext)
-	require.NoError(t, err)
-	assert.Equal(t, plaintext, decrypted)
-
-	// A second encryptor using the same dir must derive the same key (stable fallback).
-	enc2, err := newPlatformEncryptor(tmpDir)
-	require.NoError(t, err)
-	decrypted2, err := enc2.Decrypt(ciphertext)
-	require.NoError(t, err)
-	assert.Equal(t, plaintext, decrypted2, "second encryptor using same dir must decrypt ciphertext from first")
-}
-
-// TestLoadOrGenerateFallbackMachineID verifies that the fallback machine-id is stable
-// across calls and persisted to disk.
-func TestLoadOrGenerateFallbackMachineID(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("fallback machine-id is Linux-only")
-	}
-	tmpDir := t.TempDir()
-
-	id1, err := loadOrGenerateFallbackMachineID(tmpDir)
-	require.NoError(t, err)
-	require.NotEmpty(t, id1)
-
-	// Second call must return the same value (loaded from disk).
-	id2, err := loadOrGenerateFallbackMachineID(tmpDir)
-	require.NoError(t, err)
-	assert.Equal(t, id1, id2, "fallback machine-id must be stable across calls")
-
-	// Verify the file was created with restrictive permissions.
-	info, err := os.Stat(filepath.Join(tmpDir, fallbackMachineIDFile))
-	require.NoError(t, err)
-	assert.Equal(t, os.FileMode(0600), info.Mode().Perm())
-}
